@@ -1,42 +1,44 @@
 [tool.poetry]
 name = "app"
-version = "0.1.0"
-description = ""
+version = "0.0.1"
+license = "MIT"
+repository = "https://github.com/psuedomagi/ChatHoard"
+description = "ChatHoard is a ready-to-deploy full-stack web app for making the most of your ChatGPT chat history (other chat formats planned). ChatHoard offers multiple deployment options, web ui management, powerful search, and more."
 authors = ["Adam Poulemanos <mule_menthol0t at icloud.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.10"
-uvicorn = { extras = ["standard"], version = "^0.24.0.post1" }
-fastapi = "^0.109.1"
-python-multipart = "^0.0.7"
-email-validator = "^2.1.0.post1"
+python = "^3.12"
+uvicorn = { extras = ["standard"], version = "^0.30.1" }
+fastapi = "^0.111.0"
+python-multipart = "^0.0.9"
+email-validator = "^2.1.2"
 passlib = { extras = ["bcrypt"], version = "^1.7.4" }
-tenacity = "^8.2.3"
-pydantic = ">2.0"
+tenacity = "^8.4.1"
+pydantic = "^2.7.4"
 emails = "^0.6"
 
 gunicorn = "^22.0.0"
 jinja2 = "^3.1.4"
-alembic = "^1.12.1"
-httpx = "^0.25.1"
-psycopg = { extras = ["binary"], version = "^3.1.13" }
+alembic = "^1.13.1"
+httpx = "^0.27.0"
+psycopg = { extras = ["binary"], version = "^3.1.19" }
 sqlmodel = "^0.0.19"
 # Pin bcrypt until passlib supports the latest
 bcrypt = "4.0.1"
-pydantic-settings = "^2.2.1"
-sentry-sdk = { extras = ["fastapi"], version = "^1.40.6" }
+pydantic-settings = "^2.3.3"
+sentry-sdk = { extras = ["fastapi"], version = "^2.5.1" }
 pyjwt = "^2.8.0"
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^7.4.3"
-mypy = "^1.8.0"
-ruff = "^0.2.2"
-pre-commit = "^3.6.2"
+pytest = "^8.2.2"
+mypy = "^1.10.0"
+ruff = "^0.4.9"
+pre-commit = "^3.7.1"
 types-passlib = "^1.7.7.20240106"
-coverage = "^7.4.3"
+coverage = "^7.5.3"
 
 [build-system]
-requires = ["poetry>=0.12"]
+requires = ["poetry>=1.8.3"]
 build-backend = "poetry.masonry.api"
 
 [tool.mypy]
@@ -143,6 +145,10 @@ ignore = [
     "B904",   # Allow raising exceptions without from e, for HTTPException
     "COM812",
     "COM819",
+    "D100",
+    "D103",
+    "D104",
+    "D200",
     "D206",
     "D212",
     "D300",
@@ -158,6 +164,8 @@ ignore = [
     "Q001",
     "Q002",
     "Q003",
+    "S101",
+
     "UP015",
     "W191",   # indentation contains tabs
 ]
